Split runtime_tests to properly disable deprecation warnings on ComputeCpp
This moves buffer_manager_fixture to test_utils to avoid duplicating it between runtime*test.cc.
