+++ test/test_utils.h
@@ -50,7 +50,6 @@ namespace detail {
 
 namespace test_utils {
 
-
 	class mock_buffer_factory;
 
 	template <int Dims>
@@ -250,6 +249,162 @@ namespace test_utils {
 		    bid, [](int a, int b) { return a + b; }, 0, include_current_buffer_value);
 		static_cast<detail::prepass_handler&>(cgh).add_reduction<Dims>(rid);
 	}
+
+	class buffer_manager_fixture {
+	  public:
+		enum class access_target { HOST, DEVICE };
+
+		~buffer_manager_fixture() { get_device_queue().get_sycl_queue().wait_and_throw(); }
+
+		void initialize(detail::buffer_manager::buffer_lifecycle_callback cb = [](detail::buffer_manager::buffer_lifecycle_event, detail::buffer_id) {}) {
+			l = std::make_unique<detail::logger>("test", detail::log_level::warn);
+			cfg = std::make_unique<detail::config>(nullptr, nullptr, *l);
+			dq = std::make_unique<detail::device_queue>(*l);
+			dq->init(*cfg, nullptr);
+			bm = std::make_unique<detail::buffer_manager>(*dq, cb);
+			bm->enable_test_mode();
+			initialized = true;
+		}
+
+		detail::buffer_manager& get_buffer_manager() {
+			if(!initialized) initialize();
+			return *bm;
+		}
+
+		detail::device_queue& get_device_queue() {
+			if(!initialized) initialize();
+			return *dq;
+		}
+
+		static access_target get_other_target(access_target tgt) {
+			if(tgt == access_target::HOST) return access_target::DEVICE;
+			return access_target::HOST;
+		}
+
+		template <typename DataT, int Dims>
+		cl::sycl::range<Dims> get_backing_buffer_range(detail::buffer_id bid, access_target tgt, cl::sycl::range<Dims> range, cl::sycl::id<Dims> offset) {
+			if(tgt == access_target::HOST) {
+				auto info = bm->get_host_buffer<DataT, Dims>(bid, cl::sycl::access::mode::read, detail::range_cast<3>(range), detail::id_cast<3>(offset));
+				return info.buffer.get_range();
+			}
+			auto info = bm->get_device_buffer<DataT, Dims>(bid, cl::sycl::access::mode::read, detail::range_cast<3>(range), detail::id_cast<3>(offset));
+			return info.buffer.get_range();
+		}
+
+		template <typename DataT, int Dims, cl::sycl::access::mode Mode, typename KernelName = class buffer_for_each, typename Callback>
+		void buffer_for_each(detail::buffer_id bid, access_target tgt, cl::sycl::range<Dims> range, cl::sycl::id<Dims> offset, Callback cb) {
+			const auto range3 = detail::range_cast<3>(range);
+			const auto offset3 = detail::id_cast<3>(offset);
+
+			if(tgt == access_target::HOST) {
+				auto info = bm->get_host_buffer<DataT, Dims>(bid, Mode, range3, offset3);
+				const auto buf_range = detail::range_cast<3>(info.buffer.get_range());
+				for(size_t i = offset3[0]; i < offset3[0] + range3[0]; ++i) {
+					for(size_t j = offset3[1]; j < offset3[1] + range3[1]; ++j) {
+						for(size_t k = offset3[2]; k < offset3[2] + range3[2]; ++k) {
+							const auto global_idx = cl::sycl::id<3>(i, j, k);
+							const cl::sycl::id<3> local_idx = global_idx - detail::id_cast<3>(info.offset);
+							const size_t linear_idx = local_idx[0] * buf_range[1] * buf_range[2] + local_idx[1] * buf_range[2] + local_idx[2];
+							cb(detail::id_cast<Dims>(global_idx), info.buffer.get_pointer()[linear_idx]);
+						}
+					}
+				}
+			}
+
+			if(tgt == access_target::DEVICE) {
+				auto info = bm->get_device_buffer<DataT, Dims>(bid, Mode, range3, offset3);
+				const auto buf_offset = info.offset;
+				dq->get_sycl_queue()
+				    .submit([&](cl::sycl::handler& cgh) {
+					    auto acc = info.buffer.template get_access<Mode>(cgh);
+					    cgh.parallel_for<detail::bind_kernel_name<KernelName>>(range, offset, [=](cl::sycl::id<Dims> global_idx) {
+						    const auto local_idx = global_idx - buf_offset;
+						    cb(global_idx, acc[local_idx]);
+					    });
+				    })
+				    .wait();
+			}
+		}
+
+		template <typename DataT, int Dims, typename KernelName = class buffer_reduce, typename ReduceT, typename Operation>
+		ReduceT buffer_reduce(detail::buffer_id bid, access_target tgt, cl::sycl::range<Dims> range, cl::sycl::id<Dims> offset, ReduceT init, Operation op) {
+			const auto range3 = detail::range_cast<3>(range);
+			const auto offset3 = detail::id_cast<3>(offset);
+
+			if(tgt == access_target::HOST) {
+				auto info = bm->get_host_buffer<DataT, Dims>(bid, cl::sycl::access::mode::read, range3, offset3);
+				const auto buf_range = detail::range_cast<3>(info.buffer.get_range());
+				ReduceT result = init;
+				for(size_t i = offset3[0]; i < offset3[0] + range3[0]; ++i) {
+					for(size_t j = offset3[1]; j < offset3[1] + range3[1]; ++j) {
+						for(size_t k = offset3[2]; k < offset3[2] + range3[2]; ++k) {
+							const auto global_idx = cl::sycl::id<3>(i, j, k);
+							const cl::sycl::id<3> local_idx = global_idx - detail::id_cast<3>(info.offset);
+							const size_t linear_idx = local_idx[0] * buf_range[1] * buf_range[2] + local_idx[1] * buf_range[2] + local_idx[2];
+							result = op(detail::id_cast<Dims>(global_idx), result, info.buffer.get_pointer()[linear_idx]);
+						}
+					}
+				}
+				return result;
+			}
+
+			auto info = bm->get_device_buffer<DataT, Dims>(bid, cl::sycl::access::mode::read, range3, offset3);
+			const auto buf_offset = info.offset;
+			cl::sycl::buffer<ReduceT, 1> result_buf(1); // Use 1-dimensional instead of 0-dimensional since it's NYI in hipSYCL as of 0.8.1
+			// Simply do a serial reduction on the device as well
+			dq->get_sycl_queue()
+			    .submit([&](cl::sycl::handler& cgh) {
+				    auto acc = info.buffer.template get_access<cl::sycl::access::mode::read>(cgh);
+				    auto result_acc = result_buf.template get_access<cl::sycl::access::mode::read_write>(cgh);
+				    cgh.single_task<detail::bind_kernel_name<KernelName>>([=]() {
+					    result_acc[0] = init;
+					    for(size_t i = offset3[0]; i < offset3[0] + range3[0]; ++i) {
+						    for(size_t j = offset3[1]; j < offset3[1] + range3[1]; ++j) {
+							    for(size_t k = offset3[2]; k < offset3[2] + range3[2]; ++k) {
+								    const auto global_idx = cl::sycl::id<3>(i, j, k);
+								    const cl::sycl::id<3> local_idx = global_idx - detail::id_cast<3>(buf_offset);
+								    result_acc[0] = op(detail::id_cast<Dims>(global_idx), result_acc[0], acc[detail::id_cast<Dims>(local_idx)]);
+							    }
+						    }
+					    }
+				    });
+			    })
+			    .wait();
+
+			ReduceT result;
+			dq->get_sycl_queue()
+			    .submit([&](cl::sycl::handler& cgh) {
+				    auto acc = result_buf.template get_access<cl::sycl::access::mode::read>(cgh);
+				    cgh.copy(acc, &result);
+			    })
+			    .wait();
+
+			return result;
+		}
+
+		template <typename DataT, int Dims, access_mode Mode>
+		accessor<DataT, Dims, Mode, target::device> get_device_accessor(
+		    detail::live_pass_device_handler& cgh, detail::buffer_id bid, const cl::sycl::range<Dims>& range, const cl::sycl::id<Dims>& offset) {
+			auto buf_info = bm->get_device_buffer<DataT, Dims>(bid, Mode, detail::range_cast<3>(range), detail::id_cast<3>(offset));
+			return detail::make_device_accessor<DataT, Dims, Mode>(
+			    cgh.get_eventual_sycl_cgh(), subrange<Dims>(offset, range), buf_info.buffer, buf_info.offset);
+		}
+
+		template <typename DataT, int Dims, access_mode Mode>
+		accessor<DataT, Dims, Mode, target::host_task> get_host_accessor(
+		    detail::buffer_id bid, const cl::sycl::range<Dims>& range, const cl::sycl::id<Dims>& offset) {
+			auto buf_info = bm->get_host_buffer<DataT, Dims>(bid, Mode, detail::range_cast<3>(range), detail::id_cast<3>(offset));
+			return detail::make_host_accessor<DataT, Dims, Mode>(
+			    subrange<Dims>(offset, range), buf_info.buffer, buf_info.offset, detail::range_cast<Dims>(bm->get_buffer_info(bid).range));
+		}
+
+	  private:
+		bool initialized = false;
+		std::unique_ptr<detail::logger> l;
+		std::unique_ptr<detail::config> cfg;
+		std::unique_ptr<detail::device_queue> dq;
+		std::unique_ptr<detail::buffer_manager> bm;
+	};
 
 } // namespace test_utils
 } // namespace celerity
